[arcane,tests] Déplace la définition de 'MeshMaterialTesterModule' dans son propre '.h'.
